Update task 2.2.4 tool for characterisation of a given Symbol.c
switched int for uint16_t, since int can cause issues in some cases...

+++ task 2.2.4 tool for characterisation of a given Symbol.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main(){
     // this tool is supposed to identify a given Symbol
-    int toCheck;
+    uint16_t toCheck;
     // input
     printf("please input a Single Number, letter or Symbol, this Programm is gonna identify it:\n");
     scanf("%c", &toCheck);
